Make navigation font a little bit smaller
The navigation font was optimized for the custom font, but was compeltly
unreadable if the javascript files cannot be load. To avoid this issue we make
them a little bit smaller so without the custom fonts the items still fit. If
the font is loaded it's a little bit smaller but I think it's okay.