remove python requirement from main page, moving to downloads page
There was some discussion of this on mercurial@selenic.com, but with no
consensus.  Since none of the leading binary packages on the web site
(unix-like distros excepted) require a separate Python installation, I
preferred the idea of simply not mentioning the requirement at the main
page.  I've moved it to the downloads page, which allows us to explicate
a bit more as well.