downloads: put each available version into a separate table
With this update we generate a separate table on the /downloads page for
each version that is represented in sources.js.  Downloader.maxversions
is the number of versions to display by default; 0 means to display all.
If any versions are not displayed by default, a "more versions" label
appears.  Clicking it reveals the remaining versions.